scripts/requirements-audit.txt
mitreattack-python>=1.0.0
safety>=2.3.0
setuptools>=65.0.0
requests>=2.31.0
tqdm>=4.65.0 